Some users may want to extend the REQ/REP protocols similar to how REQ/REP extends XREP/XREQ.  Exposing the necessary plumbing to do that more effectively.
For a longer discussion of this change, see http://www.freelists.org/post/nanomsg/nn-req-extension
